code-review-assistant/requirements.txt
Flask==2.3.3
openai==0.28.1
werkzeug==2.3.7
python-dotenv==1.0.0
